[UI] Remove 'model' in url and breadcrumb
* Shifting the content up a level
* removes categories
so:
```
standards/model/aidc-for-patient-identification => standards/aidc-for-patient-identification
```
